went back to two clockdomains 🤷
eth_tx and eth_rx both driven by coreclk_out
for some reason now it synthesizes just fine
with vanilla liteeth

Amend:
Ahhhh, found it. The root cause of the problem is
calling add_false_path_constraint() on sys, eth_tx
and eth_rx. This adds dont_touch = true attributes
to all these nets. This is a problem for the latter
two, as they are supposed to be exchangeable with
each other and coreclk_out. This attribute makes
Vivado try to enforce these as individual nets,
with buffers between them. The solution is to use
add_false_path_constraints() with coreclk_out and
sys.